Color picker list preference (part 5):
Save clicked dialog item in onSaveInstanceState(),
so the color picker preference dialog state
can be restored after recreation.

In addition to the ListPreference state, the following states will be restored:
- Checked item in the RecyclerView
- RecyclerView scroll position, (see Note)
- Enabled state of the Dialog 'OK' button

Note:
I can't find a way to save/restore the RecyclerView scroll position
on orientation change (especially on change to landscape) which suites my needs.
Example:
- in portrait mode, there are 8 items visible
- visible item 6 is checked
- in landscape mode, there are 3 items visible
When changing to landscape, the checked item isn't visible at all in most cases,
so I decided to always make the checked item the first visible item after recreation,
(of course this is only possible, when there are enough items below the checked item,
in the example at least 7 on portrait and 2 on landscape).